High level functions for operations on MRI scans

Base.show — Method

Base.show(io::IO, z::Volume3D{T,N})

Print Volume3D info

Minc2.array — Method

array(vol::Volume3D)

Extract underlying plain array

Minc2.calculate_jacobian! — Method

calculate_jacobian!(
    tfm::Union{Vector{XFM},XFM},
    out_vol::AbstractArray{T,3},
    out_v2w::AffineTransform{C};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    ftol=1.0/80,
    max_iter=10)

Calculate dense jacobian determinant field for an arbitrary transformation

tfm - transformation to use

out_vol - output 3D array

out_v2w - voxel to world affine transform in the output array

interp - interpolation method

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.calculate_jacobian! — Method

calculate_jacobian!(
    tfm::Union{Vector{XFM},XFM}, 
    out_vol::Volume3D{T,3}; 
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    ftol=1.0/80,
    max_iter=10)::Volume3D{T,3}

Calculate dense jacobian determinant field for an arbitrary transformation

tfm - transformation to use

out_vol - output Volume3D

interp - interpolation method

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.empty_volume_like — Method

empty_volume_like(
    vol::Volume3D{T1,N}; 
    store::Type{T}=Float64, 
    history=nothing)

Create an empty Volume3D

vol - Volume3D that is used for sampling information

store - underlying array type

history - minc history

Minc2.empty_volume_like — Method

empty_volume_like(
        fn::String; 
        store::Type{T}=Float64, history=nothing)::Volume3D{T}

Create an empty Volume3D

fn - filename of a minc file that is used for sampling information

store - underlying array type

Minc2.format_history — Method

format_history(args)::String

Generate minc-style history from program args

Minc2.normalize_tfm — Method

normalize_tfm(tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::GridTransform{Float64,T}

Convert arbitrary transformation into a single GridTransform

tfm - transformation to use

ref - GridTransform that is used for sampling information

store - underlying array type

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.normalize_tfm — Method

normalize_tfm(tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::GridTransform{Float64,T}

Convert arbitrary transformation into a single GridTransform

tfm - transformation to use

ref - Volume3D that is used for sampling information

store - underlying array type

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.read_volume — Method

read_volume(fn::String; store::Type{T}=Float64)::Volume3D{T}

Read Volume3D from minc file

fn - filename

store - underlying array type

Minc2.resample_grid — Method

resample_grid(
    in_grid::Volume3D{T,4}, 
    itfm::Union{Vector{XFM}, XFM}; 
    like::Union{Nothing,Volume3D{L,4}}=nothing)::Volume3D{T,4}

Resample Volume3D that contain 4D array, using transformation, assume 1st dimension is non spatial

in_grid - input Volume3D with 4D array describing vector field

itfm - inverse of the transformation to apply (i.e from output to input)

like - Volume3D that is used for sampling information

Minc2.resample_grid_volume! — Method

resample_grid_volume!(
    in_vol::AbstractArray{T,4},
    out_vol::AbstractArray{T,4},
    v2w::AffineTransform{C}, 
    w2v::AffineTransform{C}, 
    itfm::Union{Vector{XFM}, XFM};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    fill=0.0,
    ftol=1.0/80,
    max_iter=10)::AbstractArray{T,4}

Resample 4D array using transformation, assume 1st dimension is non spatial (vector dimension)

in_vol - input 4D array

out_vol - output 4D array

v2w - voxel to world affine transform in the output array

w2v - world to voxel affine transform in the input array

itfm - inverse of the transformation to apply (i.e from output to input)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume! — Method

resample_volume!(
    in_vol::Volume3D{T,3}, 
    out_vol::Volume3D{O,3}; 
    tfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    itfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    interp::I=nothing, 
    fill=0.0, 
    order=nothing,
    ftol=1.0/80,
    max_iter=10)::Volume3D{O,3}

Resample Volume3D using transformation

in_vol - input Volume3D

out_vol - output Volume3D

itfm - inverse of the transformation to apply (i.e from output to input)

tfm - transformation to apply (i.e from output to input) (instead of itfm)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume! — Method

resample_volume!(in_vol::AbstractArray{T,3}, 
    out_vol::AbstractArray{T,3}, 
    v2w::AffineTransform{C}, 
    w2v::AffineTransform{C}, 
    itfm::Union{Vector{XFM},XFM};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    fill=0.0,
    ftol=1.0/80,
    max_iter=10)

Resample 3D array using transformation

in_vol - input 3D array

out_vol - output 3D array

v2w - voxel to world affine transform in the output array

w2v - world to voxel affine transform in the input array

itfm - inverse of the transformation to apply (i.e from output to input)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume — Method

resample_volume(
    in_vol::Volume3D{T,3};
    like::Union{Volume3D{O,3},Nothing}=nothing,
    tfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    itfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    interp::I=nothing, 
    fill=0.0,
    order=1,
    ftol=1.0/80,
    max_iter=10)::Volume3D

Resample Volume3D using transformation

in_vol - input Volume3D

like - Volume3D that is used for sampling information

itfm - inverse of the transformation to apply (i.e from output to input)

tfm - transformation to apply (i.e from output to input) (instead of itfm)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.reshape_volume — Method

Reshape Volume3D Work in progress DO NOT USE

Minc2.save_volume — Method

save_volume(fn::AbstractString, 
    vol::Volume3D{T,N}; 
    store::Type{S}=Float32,
    history=nothing)

Save Volume3D to minc file

fn - filename

vol - Volume3D to save

store - underlying MINC data type, to be used for storage

Minc2.tfm_to_grid! — Method

tfm_to_grid!(
    tfm::Union{Vector{XFM}, XFM}, 
    grid::AbstractArray{T,4},
    v2w::AffineTransform{C};
    ftol=1.0/80,max_iter=10)::AbstractArray{T,4}

Convert arbitrary transformation into vector field

tfm - transformation to use

grid - output 4D array, will contain vector field

v2w - voxel to world affine transform in the output array

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.tfm_to_grid — Method

tfm_to_grid(
    tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::Volume3D{T,4}

Convert arbitrary transformation into Volume3D with 4D array containing vector field

tfm - transformation to use

v2w - voxel to world affine transform in the output array

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.voxel_to_world — Method

voxel_to_world(vol::Volume3D)

Extract voxel to world affine transform from a Volume3D

Minc2.world_to_voxel — Method

world_to_voxel(vol::Volume3D)

Extract world to voxel affine transform from a Volume3D

Minc2.GridTransform — Method

GridTransform(vol::V)

Create GridTransform from Volume3D, assume it is a vector field

vol - Volume3D with 4D array containing vector field

Minc2.InverseGridTransform — Method

InverseGridTransform(vol::V)

Create InverseGridTransform from Volume3D, assume it is a vector field

vol - Volume3D with 4D array containing vector field

Minc2.Volume3D — Type

Volume3D{T,N}

An abstract 3D volume, could be vector field or volume with time dimension

Minc2.Volume3D — Method

Volume3D(vol::Array{T,N}, v2w::AffineTransform{Float64}; 
    history::Union{AbstractString,Nothing}=nothing)::Volume3D{T,N}

Create Volume3D from an array and affine transform

vol - underlying array

v2w - voxel to world affine transform

history - minc history

Minc2.Volume3D — Method

Volume3D(vol::Array{T,N}, like::Volume3D; 
    history::Union{AbstractString,Nothing}=nothing)::Volume3D{T,N}

Create Volume3D from an array and another Volume3D that is used for sampling information

vol - underlying array

like - Volume3D that is used for sampling information

history - minc history

Minc2.draw_outline_with_heatmap — Method

draw_outline_with_heatmap(
    layout, anat, heat; 
    cmap=:rainbow, levels=[20, 40],
    heat_limits=nothing, 
    show_colorbar=true,
    nslices=4)

Draws a heatmap of heat on top of anat in a grid layout layout.

layout is a grid layout

anat is a Volume3D with anatomical image

heat is a Volume3D with overlay heatmap

heat_limits limits for the heatmap

cmap is a color map to use

levels is a list of levels of anat to use for contours

show_colorbar is a boolean flag to show colorbar

nslices is a number of slices to show

Minc2.draw_outline_with_labels — Method

draw_outline_with_labels(
        layout, 
        anat, 
        seg; 
        labels, 
        cmap = :rainbow,
        levels = [20,30],
        show_colorbar = true,
        nslices = 4)

Draws a segmentation seg on top of anat in a grid layout layout

layout is a grid layout

anat is a Volume3D with anatomical image

seg is a Volume3D with discrete segemntations

labels is a dictionary of labels to show

cmap is a color map to use

levels is a list of levels of anat to use for contours

show_colorbar is a boolean flag to show colorbar

nslices is a number of slices to show

Minc2.lps_to_ras — Method

lps_to_ras(xfm::Minc2.AffineTransform)

Convert coordinates from LPS to RAS

Minc2.lps_to_ras — Method

lps_to_ras(vol::Minc2.Volume3D{T})::Minc2.Volume3D{T}

Convert coordinates from LPS to RAS

Minc2.read_ants_transform — Method

read_ants_transform(fn::AbstractString; store::Type{T}=Float32)::Minc2.AnyTransform

Read .txt and .nii(.nii.gz) transforms produces by ANTs

Minc2.read_itk_nifti_transform — Method

read_itk_nifti_transform(fn::AbstractString;
    store::Type{T}=Float32)::Minc2.GridTransform{Float64,T}

Read ANTs style warp transform

Minc2.read_itk_txt_transform — Method

read_itk_txt_transform(fn::AbstractString)::Minc2.AffineTransform{Float64}

Read ITK legacy transform in .txt format

Minc2.read_nifti_volume — Method

read_nifti_volume(fn::AbstractString; store::Type{T}=Float64)::Volume3D{T}

Read Volume3D from .nii or .nii.gz file

Minc2.save_nifti_volume — Method

save_nifti_volume(fn::AbstractString, vol::Volume3D{T}; 
    store::Type{S}=Float32, history=nothing)

Save Volume3D into .nii or .nii.gz file

Minc2.write_itk_nifti_transform — Method

write_itk_nifti_transform(fn::AbstractString,
    xfm::Minc2.GridTransform{Float64,T}; store::Type{S}=Float32)

Write ANTs style warp transform

Minc2.write_itk_txt_transform — Method

write_itk_txt_transform(fn::AbstractString,
    xfm::Minc2.AffineTransform{Float64})

Write affine transform in ITK .txt format