Low level functions for IO on .mnc and .xfm files

Minc2.append_grid_transform — Method

append_grid_transform(h::TransformHandle, grid_file::String; inv::Bool=false)

Low level: Append grid transform to an open transformation handle

Minc2.append_linear_transform — Method

append_linear_transform(h::TransformHandle, lin::AffineTransform)

Low level: Append affine transform to an open transformation handle

Minc2.concat_xfm — Method

concat_xfm(h::TransformHandle, i::TransformHandle)

Low level: concatenate two transfomations

Minc2.get_grid_transform — Method

get_grid_transform(h::TransformHandle; n::Int64=0)

Low level: extract reference to a grid file from open handle

Minc2.get_linear_transform — Method

get_linear_transform(h::TransformHandle; n::Int64=0)::AffineTransform{Float64}

Low level: extract AffineTransform{Float64} from open handle

Minc2.get_linear_transform_param — Method

get_linear_transform_param(h::TransformHandle;n::Int64=0,center::Union{Nothing,Vector{Float64}}=nothing)

Low level: extact transformation parameters from affine transform , returns named tuple with fields center, translations, scales, shears, rotations

Minc2.get_n_concat — Method

get_n_concat(h::TransformHandle)::Int64

Low level: Get number of transformations in open handle

Minc2.get_n_type — Method

get_n_type(h::TransformHandle; n::Int64=0)::XFM

Low level: Get transform type for nth transform in open handle

Minc2.inverse_transform_point — Method

inverse_transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Inverse transform point xyz using minc transform h

Minc2.invert_transform — Method

invert_transform(h::TransformHandle)

Low level: Invert minc transform h

Minc2.load_transforms — Method

load_transforms(h::TransformHandle)::Vector{AnyTransform}

Low level: Load all transforms from open .XFM handle

Minc2.load_transforms — Method

load_transforms(fname::String)::Vector{AnyTransform}

Load transformations from .xfm file fname

Minc2.open_xfm_file — Method

Low level: Open transform xfm file, return handle

Minc2.read_transforms — Method

read_transforms(fname::String)::Vector{AnyTransform}

Read transformations from .xfm file fname

Minc2.save_transforms — Method

save_transforms(fname::String, 
    xfm::Union{Vector{XFM}, XFM};
    grid_store::Type{T}=Float32 ) where {T, XFM<:AnyTransform}

Save transformations into .xfm file,

Arguments

fname: output file name

xfm: vector of transformations to save, or a single transformation

grid_store: type of storage for grid files (default: Float32)

Minc2.save_xfm_file — Method

save_xfm_file(h::TransformHandle, path::String)

Low level: Save information into file from an open handle

Minc2.transform_point — Method

transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Transform point xyz using minc transform h

Minc2.TransformHandle — Type

Low level: minc2_simple XFM transform handle