Low level functions for IO on .mnc and .xfm files

Minc2.@minc2_check — Macro

minc2_check( ex )

Macro to verify the return code

Minc2._hdr_convert! — Method

_hdr_convert!(hdr::MincHeader,dd)::MincHeader

internal function

Minc2.attributes — Method

attributes(h::VolumeHandle, g::String)

List attributes defined inside given group

Minc2.close_minc_file — Method

close_minc_file(h::VolumeHandle)

Close currently open minc file, commit data on disk

Minc2.copy_minc_metadata — Method

copy_minc_metadata(i::VolumeHandle, o::VolumeHandle)

Copy metadata from one file to another

Minc2.create_header_from_v2w — Method

create_header_from_v2w(
    sz, t::AffineTransform{T};
    vector_dim::Bool=false, 
    time_step::Union{Float64,Nothing}=nothing,
    time_start::Union{Float64,Nothing}=nothing)::MincHeader

Internal: Generate header from the voxel to world transform and volume size

Minc2.create_minc_file — Method

create_minc_file(h::VolumeHandle, path::AbstractString)

Create empty minc file on disk, structure need to be defined in advance

Minc2.define_minc_file — Method

define_minc_file(hdr::MincHeader,::Type{Store}=Float32,::Type{Repr}=Store)

Define new minc file structure, return handle

Minc2.empty_like_minc_volume_raw — Method

empty_like_minc_volume_raw( h::VolumeHandle,
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader}

Allocate empty volume using handle return volume, storage header

Minc2.empty_like_minc_volume_raw — Method

empty_like_minc_volume_raw(path::String, ::Type{T}=Float32 )::
        Tuple{Array{T}, Minc2.MincHeader}

Create empty volume similar to existing file return volume, representation header,storage header

Minc2.empty_like_minc_volume_std — Method

empty_like_minc_volume_std(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using handle return volume, representation header,storage header

Minc2.empty_like_minc_volume_std — Method

empty_like_minc_volume_std(path::String, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

allocate empty volume using path as a reference return volume, representation header,storage header

Minc2.empty_like_minc_volume_std_history — Method

empty_like_minc_volume_std_history(path::String, ::Type{T}=Float32 )

allocate empty volume using path return volume, representation header,storage header

Minc2.get_attribute — Method

get_attribute(h::VolumeHandle,g::String,a::String; default=missing)

Convenience function for reading specific attribute, return default value if not found also convert Array into the first value if it's a one-length array

Minc2.groups — Method

groups(h::VolumeHandle)

List groups defined in minc2 file

Minc2.ndim — Method

ndim(h::VolumeHandle)::Int

Query number of dimensions in the minc file

Minc2.open_minc_file — Method

open_minc_file(fname::String)::VolumeHandle

Open minc file, return handle

Minc2.read_attribute — Method

read_attribute(h::VolumeHandle, 
    group::String,
    attribute::String)::Union{String, AbstractVector, Nothing}

Read minc2 header attribute :param group: attribute group name :param attribute: attribute name :return:

Minc2.read_history — Method

read_history(i::VolumeHandle)::Union{String, Nothing}

Return history string

Minc2.read_minc_volume_raw — Method

read_minc_volume_raw(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader}

Read the actual volume using handle return volume, storage header

Minc2.read_minc_volume_raw — Method

read_minc_volume_raw(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader}

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_raw_history — Method

read_minc_volume_raw_history(path::String, ::Type{T}=Float32 )

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_std — Method

read_minc_volume_std(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using handle return volume, representation header,storage header

Minc2.read_minc_volume_std — Method

read_minc_volume_std(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_std_history — Method

read_minc_volume_std_history(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader, Union{String,Nothing}}

Read the actual volume using path return volume, representation header,storage header

Minc2.representation_header — Method

representation_header(h::VolumeHandle)::MincHeader

Return volume representation header

Minc2.setup_standard_order — Method

setup_standard_order(h::VolumeHandle)

Prepare to read volume in standard order: [V,X,Y,Z,TIME]

Minc2.store_header — Method

store_header(h::VolumeHandle)::MincHeader

Return volume on-disk stucture header

Minc2.voxel_to_world — Method

voxel_to_world(hdr::MincHeader)::AffineTransform{Float64}

Give AffineTransform for world to voxel transformation based on header

Minc2.voxel_to_world — Method

voxel_to_world(h::VolumeHandle,ijk::Vector{Float64})::Vector{Float64}

Convert contignuous 0-based voxel indexes (I,J,K) to world coordinates (X,Y,Z) 0-based

Minc2.world_to_voxel — Method

world_to_voxel(hdr::MincHeader)::AffineTransform{Float64}

Give AffineTransform for voxel to world transformation

Minc2.world_to_voxel — Method

world_to_voxel(h::VolumeHandle, xyz::Vector{Float64})::Vector{Float64}

Convert world coordinates (X,Y,Z) to contignuous voxel indexes (I,J,K) 0-based

Minc2.write_attribute — Method

write_attribute(h::VolumeHandle, group::String, attribute::String, value::T)

Store attribute into minc2 file :param group: group name :param attribute: attribute name :param value: attribute value :return:

Minc2.write_history — Method

write_history(i::VolumeHandle, history::String)

Write history string

Minc2.write_minc_volume_raw — Method

write_minc_volume_raw(path::String, ::Type{Store}, 
    store_hdr::Union{MincHeader, Nothing}, 
    volume::Array{Repr};like::Union{String, Nothing}=nothing, 
    history::Union{String, Nothing}=nothing )

Write full volume to file, need to provide details of file structure return nothing

Minc2.write_minc_volume_raw — Method

write_minc_volume_raw(h::VolumeHandle, volume::Array{T} )

write full volume to file, file should be defined and created return nothing

Minc2.write_minc_volume_std — Method

write_minc_volume_std(path::String, ::Type{Store}, 
    store_hdr::Union{MincHeader, Nothing}, 
    volume::Array{Repr};like::Union{String, Nothing}=nothing, 
    history::Union{String, Nothing}=nothing )

Write full volume to file, need to provide details of file structure return nothing

Minc2.write_minc_volume_std — Method

write_minc_volume_std(h::VolumeHandle, volume::Array{T} )

write full volume to file, file should be defined and created return nothing

Minc2.DIM — Type

Low Level: Axis types from MINC volume

Minc2.Minc2Error — Type

MINC io exception

Minc2.MincHeader — Type

MincHeader

Structure describing spatial orientation and sampling of the minc file

Minc2.STATUS — Type

Low Level: minc2_simple API status

Minc2.VolumeHandle — Type

VolumeHandle

minc2_simple volume handle

Minc2.julia_to_minc2 — Constant

Low Level: map Julia types to minc2 data types

Minc2.minc2_spatial — Constant

Low Level: Mapping MINC2 spatial dimensions to proper spatial dimension (should be identity for spatial dims)

Minc2.minc2_to_julia — Constant

Low Level: map MINC2 types to Julia

Minc2.append_grid_transform — Method

append_grid_transform(h::TransformHandle, grid_file::String; inv::Bool=false)

Low level: Append grid transform to an open transformation handle

Minc2.append_linear_transform — Method

append_linear_transform(h::TransformHandle, lin::AffineTransform)

Low level: Append affine transform to an open transformation handle

Minc2.concat_xfm — Method

concat_xfm(h::TransformHandle, i::TransformHandle)

Low level: concatenate two transfomations

Minc2.get_grid_transform — Method

get_grid_transform(h::TransformHandle; n::Int64=0)

Low level: extract reference to a grid file from open handle

Minc2.get_linear_transform — Method

get_linear_transform(h::TransformHandle; n::Int64=0)::AffineTransform{Float64}

Low level: extract AffineTransform{Float64} from open handle

Minc2.get_linear_transform_param — Method

get_linear_transform_param(h::TransformHandle;n::Int64=0,center::Union{Nothing,Vector{Float64}}=nothing)

Low level: extact transformation parameters from affine transform , returns named tuple with fields center, translations, scales, shears, rotations

Minc2.get_n_concat — Method

get_n_concat(h::TransformHandle)::Int64

Low level: Get number of transformations in open handle

Minc2.get_n_type — Method

get_n_type(h::TransformHandle; n::Int64=0)::XFM

Low level: Get transform type for nth transform in open handle

Minc2.inverse_transform_point — Method

inverse_transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Inverse transform point xyz using minc transform h

Minc2.invert_transform — Method

invert_transform(h::TransformHandle)

Low level: Invert minc transform h

Minc2.load_transforms — Method

load_transforms(h::TransformHandle)::Vector{AnyTransform}

Low level: Load all transforms from open .XFM handle

Minc2.load_transforms — Method

load_transforms(fname::String)::Vector{AnyTransform}

Load transformations from .xfm file fname

Minc2.open_xfm_file — Method

Low level: Open transform xfm file, return handle

Minc2.read_transforms — Method

read_transforms(fname::String)::Vector{AnyTransform}

Read transformations from .xfm file fname

Minc2.save_transforms — Method

save_transforms(fname::String, 
    xfm::Union{Vector{XFM}, XFM};
    grid_store::Type{T}=Float32 ) where {T, XFM<:AnyTransform}

Save transformations into .xfm file,

Arguments

fname: output file name

xfm: vector of transformations to save, or a single transformation

grid_store: type of storage for grid files (default: Float32)

Minc2.save_xfm_file — Method

save_xfm_file(h::TransformHandle, path::String)

Low level: Save information into file from an open handle

Minc2.transform_point — Method

transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Transform point xyz using minc transform h

Minc2.TransformHandle — Type

Low level: minc2_simple XFM transform handle

Minc2.XFM — Type

@enum XFM

Low level: transformation types stored in .xfm file