Base.length — Method

Base.length(tr::AnyTransform)

Length of concatenated transform

Base.show — Method

Base.show(io::IO, z::GridTransform{T,F,I})

Print summary information about grid transform

Base.show — Method

Base.show(io::IO, z::InverseGridTransform{T,F,I})

Print summary information about inverse grid transform

Minc2.array — Method

array(grid::GridTransform)

Extract underlying plain array

grid - GridTransform

Minc2.array — Method

array(grid::InverseGridTransform)

Extract underlying plain array

grid - InverseGridTransform

Minc2.decompose — Method

decompose(rot, shift)

Decompose affine transform specified as rotation matrix ans shift vector into three components start, step, direction cosines

rot - rotation matrix (3x3)

shift - shift vector (3)

Minc2.decompose — Method

decompose(tfm::Matrix{T})

Decompose affine transform into three components start, step, direction cosines

Minc2.decompose — Method

decompose(tfm::AffineTransform{T})

Decompose affine transform into three components start, step, direction cosines

tfm - affine transform

Minc2.interpolate_field — Method

interpolate_field(
        v2w::AffineTransform{T},
        itp_vector_field::I, 
        p::SVector{3,T} )::SVector{3,T}

Internal support function

v2w - voxel to world affine transform

itp_vector_field - vector field

p - point to transform

Minc2.inv — Method

inv(::IdentityTransform)::IdentityTransform

Invert IdentityTransform transform, does nothing

Minc2.inv — Method

inv(t::AffineTransform{T})::AffineTransform{T}

Invert AffineTransform transform

t - AffineTransform

Minc2.inv — Method

inv(t::GridTransform{T,F,VF})::InverseGridTransform{T,F,VF}

Invert GridTransform transform

t - GridTransform

Minc2.inv — Method

inv(t::InverseGridTransform{T,F,VF})::GridTransform{T,F,VF}

Invert InverseGridTransform transform

t - InverseGridTransform

Minc2.inv — Method

inv(t::Vector{T})::Vector{AnyTransform}

Invert concatenated transform

t - concatenated transform

Minc2.transform_point — Method

transform_point(
        tfm::GridTransform{T,F}, p::SVector{3,T};
        _whatever...)::SVector{3,T} where {T,F}

Apply forward grid transform to a point

tfm - grid transform

p - point to transform

Minc2.transform_point — Method

transform_point(
    tfm::AffineTransform{T}, 
    p::CartesianIndex{3};
    _whatever...)::SVector{3,T}

Apply inverse grid transform reimplements algorithm from MNIformats/gridtransforms.c:gridinversetransform_point

tfm - affine transform

p - point to transform

Minc2.transform_point — Method

transform_point(
    tfm::AffineTransform{T}, 
    p::CartesianIndex{3};
    _whatever...)::SVector{3,T}

Apply affine transform to CartesianIndices

tfm - affine transform

p - point to transform

Minc2.transform_point — Method

transform_point(
        tfm::AffineTransform{T}, 
        p::SVector{3,T};
        _whatever...)::SVector{3,T}

Apply affine transform to a point

tfm - affine transform

p - point to transform

Minc2.transform_point — Method

transform_point(tfm::IdentityTransform, 
    p::SVector{3,T};
    _whatever...)

tfm - IdentityTransform

p - point to transform

Apply identity transform to a point (returns same point)

Minc2.transform_point — Method

transform_point(
    tfm::Vector{XFM},
    p::SVector{3,T};
    max_iter::Int=10,
    ftol::Float64=1.0/80)::SVector{3,T}

Apply concatenated transform to a point

tfm - concatenated transform

p - point to transform

max_iter - maximum number of iterations for inverse transform

ftol - tolerance for inverse transform

Minc2.voxel_to_world — Method

voxel_to_world(grid::GridTransform)

Extract voxel to world affine transform from a GridTransform

grid - GridTransform

Minc2.voxel_to_world — Method

voxel_to_world(grid::InverseGridTransform)

Extract voxel to world affine transform from a InverseGridTransform

grid - InverseGridTransform

Minc2.world_to_voxel — Method

world_to_voxel(grid::GridTransform)

Extract world to voxel affine transform from a GridTransform

grid - GridTransform

Minc2.world_to_voxel — Method

world_to_voxel(grid::InverseGridTransform)

Extract world to voxel affine transform from a InverseGridTransform

grid - InverseGridTransform

Minc2.AffineTransform — Type

AffineTransform{T}

Affine transform, described by rotation matrix and shift vector

Minc2.AffineTransform — Method

AffineTransform(rot, shift) Create affine transform from rotation and shift

rot - rotation matrix-like (3x3)

shift - shift vector-like (3)

Minc2.AffineTransform — Method

AffineTransform(mat)  Create affine transform from a matrix-like object

mat - matrix (4x4 or 3x4)

Minc2.AnyTransform — Type

AnyTransform()

Any 3D Geometrical transform, abstract type

Minc2.GeoTransforms — Type

GeoTransforms=Vector{AnyTransform}

Concatenated transforms

Minc2.GridTransform — Type

GridTransform{T,F,I}

Dense vector field transform (grid transform)

Minc2.GridTransform — Method

GridTransform(
    voxel_to_world::AffineTransform{T},
    vector_field::Array{F, 4})

Construct GridTransform from voxel to world transform and a vector field

voxel_to_world - voxel to world affine transform

vector_field - vector field

Minc2.GridTransform — Method

GridTransform(::Type{T}=Float64,::Type{F}=Float64)

Construct empty GridTransform, which should generate identity transform

Minc2.IdentityTransform — Type

IdentityTransform()

Identity transform, does nothing

Minc2.InverseGridTransform — Type

InverseGridTransform{T,F,VF}

Dense vector field transform (GridTransform) used in inverse

Minc2.InverseGridTransform — Method

InverseGridTransform(
    voxel_to_world::AffineTransform{T},
    vector_field::Array{F, 4})

Construct InverseGridTransform from voxel to world transform and a vector field

voxel_to_world - voxel to world affine transform

vector_field - vector field

Minc2.InverseGridTransform — Method

InverseGridTransform(::Type{T}=Float64,::Type{F}=Float64)

Construct InverseGridTransform empty transform

Minc2.@minc2_check — Macro

minc2_check( ex )

Macro to verify the return code

Minc2._hdr_convert! — Method

_hdr_convert!(hdr::MincHeader,dd)::MincHeader

internal function

Minc2.attributes — Method

attributes(h::VolumeHandle, g::String)

List attributes defined inside given group

Minc2.close_minc_file — Method

close_minc_file(h::VolumeHandle)

Close currently open minc file, commit data on disk

Minc2.copy_minc_metadata — Method

copy_minc_metadata(i::VolumeHandle, o::VolumeHandle)

Copy metadata from one file to another

Minc2.create_header_from_v2w — Method

create_header_from_v2w(
    sz, t::AffineTransform{T};
    vector_dim::Bool=false, 
    time_step::Union{Float64,Nothing}=nothing,
    time_start::Union{Float64,Nothing}=nothing)::MincHeader

Internal: Generate header from the voxel to world transform and volume size

Minc2.create_minc_file — Method

create_minc_file(h::VolumeHandle, path::AbstractString)

Create empty minc file on disk, structure need to be defined in advance

Minc2.define_minc_file — Method

define_minc_file(hdr::MincHeader,::Type{Store}=Float32,::Type{Repr}=Store)

Define new minc file structure, return handle

Minc2.empty_like_minc_volume_raw — Method

empty_like_minc_volume_raw( h::VolumeHandle,
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader}

Allocate empty volume using handle return volume, storage header

Minc2.empty_like_minc_volume_raw — Method

empty_like_minc_volume_raw(path::String, ::Type{T}=Float32 )::
        Tuple{Array{T}, Minc2.MincHeader}

Create empty volume similar to existing file return volume, representation header,storage header

Minc2.empty_like_minc_volume_std — Method

empty_like_minc_volume_std(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using handle return volume, representation header,storage header

Minc2.empty_like_minc_volume_std — Method

empty_like_minc_volume_std(path::String, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

allocate empty volume using path as a reference return volume, representation header,storage header

Minc2.empty_like_minc_volume_std_history — Method

empty_like_minc_volume_std_history(path::String, ::Type{T}=Float32 )

allocate empty volume using path return volume, representation header,storage header

Minc2.get_attribute — Method

get_attribute(h::VolumeHandle,g::String,a::String; default=missing)

Convenience function for reading specific attribute, return default value if not found also convert Array into the first value if it's a one-length array

Minc2.groups — Method

groups(h::VolumeHandle)

List groups defined in minc2 file

Minc2.ndim — Method

ndim(h::VolumeHandle)::Int

Query number of dimensions in the minc file

Minc2.open_minc_file — Method

open_minc_file(fname::String)::VolumeHandle

Open minc file, return handle

Minc2.read_attribute — Method

read_attribute(h::VolumeHandle, 
    group::String,
    attribute::String)::Union{String, AbstractVector, Nothing}

Read minc2 header attribute :param group: attribute group name :param attribute: attribute name :return:

Minc2.read_history — Method

read_history(i::VolumeHandle)::Union{String, Nothing}

Return history string

Minc2.read_minc_volume_raw — Method

read_minc_volume_raw(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader}

Read the actual volume using handle return volume, storage header

Minc2.read_minc_volume_raw — Method

read_minc_volume_raw(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader}

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_raw_history — Method

read_minc_volume_raw_history(path::String, ::Type{T}=Float32 )

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_std — Method

read_minc_volume_std(h::VolumeHandle, 
    ::Type{T}=Float32 )::Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using handle return volume, representation header,storage header

Minc2.read_minc_volume_std — Method

read_minc_volume_std(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader}

Read the actual volume using path return volume, representation header,storage header

Minc2.read_minc_volume_std_history — Method

read_minc_volume_std_history(path::String, ::Type{T}=Float32 )::
    Tuple{Array{T}, Minc2.MincHeader, Minc2.MincHeader, Union{String,Nothing}}

Read the actual volume using path return volume, representation header,storage header

Minc2.representation_header — Method

representation_header(h::VolumeHandle)::MincHeader

Return volume representation header

Minc2.setup_standard_order — Method

setup_standard_order(h::VolumeHandle)

Prepare to read volume in standard order: [V,X,Y,Z,TIME]

Minc2.store_header — Method

store_header(h::VolumeHandle)::MincHeader

Return volume on-disk stucture header

Minc2.voxel_to_world — Method

voxel_to_world(hdr::MincHeader)::AffineTransform{Float64}

Give AffineTransform for world to voxel transformation based on header

Minc2.voxel_to_world — Method

voxel_to_world(h::VolumeHandle,ijk::Vector{Float64})::Vector{Float64}

Convert contignuous 0-based voxel indexes (I,J,K) to world coordinates (X,Y,Z) 0-based

Minc2.world_to_voxel — Method

world_to_voxel(hdr::MincHeader)::AffineTransform{Float64}

Give AffineTransform for voxel to world transformation

Minc2.world_to_voxel — Method

world_to_voxel(h::VolumeHandle, xyz::Vector{Float64})::Vector{Float64}

Convert world coordinates (X,Y,Z) to contignuous voxel indexes (I,J,K) 0-based

Minc2.write_attribute — Method

write_attribute(h::VolumeHandle, group::String, attribute::String, value::T)

Store attribute into minc2 file :param group: group name :param attribute: attribute name :param value: attribute value :return:

Minc2.write_history — Method

write_history(i::VolumeHandle, history::String)

Write history string

Minc2.write_minc_volume_raw — Method

write_minc_volume_raw(path::String, ::Type{Store}, 
    store_hdr::Union{MincHeader, Nothing}, 
    volume::Array{Repr};like::Union{String, Nothing}=nothing, 
    history::Union{String, Nothing}=nothing )

Write full volume to file, need to provide details of file structure return nothing

Minc2.write_minc_volume_raw — Method

write_minc_volume_raw(h::VolumeHandle, volume::Array{T} )

write full volume to file, file should be defined and created return nothing

Minc2.write_minc_volume_std — Method

write_minc_volume_std(path::String, ::Type{Store}, 
    store_hdr::Union{MincHeader, Nothing}, 
    volume::Array{Repr};like::Union{String, Nothing}=nothing, 
    history::Union{String, Nothing}=nothing )

Write full volume to file, need to provide details of file structure return nothing

Minc2.write_minc_volume_std — Method

write_minc_volume_std(h::VolumeHandle, volume::Array{T} )

write full volume to file, file should be defined and created return nothing

Minc2.DIM — Type

Low Level: Axis types from MINC volume

Minc2.Minc2Error — Type

MINC io exception

Minc2.MincHeader — Type

MincHeader

Structure describing spatial orientation and sampling of the minc file

Minc2.STATUS — Type

Low Level: minc2_simple API status

Minc2.VolumeHandle — Type

VolumeHandle

minc2_simple volume handle

Minc2.julia_to_minc2 — Constant

Low Level: map Julia types to minc2 data types

Minc2.minc2_spatial — Constant

Low Level: Mapping MINC2 spatial dimensions to proper spatial dimension (should be identity for spatial dims)

Minc2.minc2_to_julia — Constant

Low Level: map MINC2 types to Julia

Base.show — Method

Base.show(io::IO, z::Volume3D{T,N})

Print Volume3D info

Minc2.array — Method

array(vol::Volume3D)

Extract underlying plain array

Minc2.calculate_jacobian! — Method

calculate_jacobian!(
    tfm::Union{Vector{XFM},XFM},
    out_vol::AbstractArray{T,3},
    out_v2w::AffineTransform{C};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    ftol=1.0/80,
    max_iter=10)

Calculate dense jacobian determinant field for an arbitrary transformation

tfm - transformation to use

out_vol - output 3D array

out_v2w - voxel to world affine transform in the output array

interp - interpolation method

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.calculate_jacobian! — Method

calculate_jacobian!(
    tfm::Union{Vector{XFM},XFM}, 
    out_vol::Volume3D{T,3}; 
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    ftol=1.0/80,
    max_iter=10)::Volume3D{T,3}

Calculate dense jacobian determinant field for an arbitrary transformation

tfm - transformation to use

out_vol - output Volume3D

interp - interpolation method

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.empty_volume_like — Method

empty_volume_like(
    vol::Volume3D{T1,N}; 
    store::Type{T}=Float64, 
    history=nothing)

Create an empty Volume3D

vol - Volume3D that is used for sampling information

store - underlying array type

history - minc history

Minc2.empty_volume_like — Method

empty_volume_like(
        fn::String; 
        store::Type{T}=Float64, history=nothing)::Volume3D{T}

Create an empty Volume3D

fn - filename of a minc file that is used for sampling information

store - underlying array type

Minc2.format_history — Method

format_history(args)::String

Generate minc-style history from program args

Minc2.normalize_tfm — Method

normalize_tfm(tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::GridTransform{Float64,T}

Convert arbitrary transformation into a single GridTransform

tfm - transformation to use

ref - GridTransform that is used for sampling information

store - underlying array type

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.normalize_tfm — Method

normalize_tfm(tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::GridTransform{Float64,T}

Convert arbitrary transformation into a single GridTransform

tfm - transformation to use

ref - Volume3D that is used for sampling information

store - underlying array type

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.read_volume — Method

read_volume(fn::String; store::Type{T}=Float64)::Volume3D{T}

Read Volume3D from minc file

fn - filename

store - underlying array type

Minc2.resample_grid — Method

resample_grid(
    in_grid::Volume3D{T,4}, 
    itfm::Union{Vector{XFM}, XFM}; 
    like::Union{Nothing,Volume3D{L,4}}=nothing)::Volume3D{T,4}

Resample Volume3D that contain 4D array, using transformation, assume 1st dimension is non spatial

in_grid - input Volume3D with 4D array describing vector field

itfm - inverse of the transformation to apply (i.e from output to input)

like - Volume3D that is used for sampling information

Minc2.resample_grid_volume! — Method

resample_grid_volume!(
    in_vol::AbstractArray{T,4},
    out_vol::AbstractArray{T,4},
    v2w::AffineTransform{C}, 
    w2v::AffineTransform{C}, 
    itfm::Union{Vector{XFM}, XFM};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    fill=0.0,
    ftol=1.0/80,
    max_iter=10)::AbstractArray{T,4}

Resample 4D array using transformation, assume 1st dimension is non spatial (vector dimension)

in_vol - input 4D array

out_vol - output 4D array

v2w - voxel to world affine transform in the output array

w2v - world to voxel affine transform in the input array

itfm - inverse of the transformation to apply (i.e from output to input)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume! — Method

resample_volume!(
    in_vol::Volume3D{T,3}, 
    out_vol::Volume3D{O,3}; 
    tfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    itfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    interp::I=nothing, 
    fill=0.0, 
    order=nothing,
    ftol=1.0/80,
    max_iter=10)::Volume3D{O,3}

Resample Volume3D using transformation

in_vol - input Volume3D

out_vol - output Volume3D

itfm - inverse of the transformation to apply (i.e from output to input)

tfm - transformation to apply (i.e from output to input) (instead of itfm)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume! — Method

resample_volume!(in_vol::AbstractArray{T,3}, 
    out_vol::AbstractArray{T,3}, 
    v2w::AffineTransform{C}, 
    w2v::AffineTransform{C}, 
    itfm::Union{Vector{XFM},XFM};
    interp::I=BSpline(Quadratic(Line(OnCell()))),
    fill=0.0,
    ftol=1.0/80,
    max_iter=10)

Resample 3D array using transformation

in_vol - input 3D array

out_vol - output 3D array

v2w - voxel to world affine transform in the output array

w2v - world to voxel affine transform in the input array

itfm - inverse of the transformation to apply (i.e from output to input)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.resample_volume — Method

resample_volume(
    in_vol::Volume3D{T,3};
    like::Union{Volume3D{O,3},Nothing}=nothing,
    tfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    itfm::Union{Vector{XFM},XFM,Nothing}=nothing, 
    interp::I=nothing, 
    fill=0.0,
    order=1,
    ftol=1.0/80,
    max_iter=10)::Volume3D

Resample Volume3D using transformation

in_vol - input Volume3D

like - Volume3D that is used for sampling information

itfm - inverse of the transformation to apply (i.e from output to input)

tfm - transformation to apply (i.e from output to input) (instead of itfm)

interp - interpolation method

fill - fill value

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.reshape_volume — Method

Reshape Volume3D Work in progress DO NOT USE

Minc2.save_volume — Method

save_volume(fn::AbstractString, 
    vol::Volume3D{T,N}; 
    store::Type{S}=Float32,
    history=nothing)

Save Volume3D to minc file

fn - filename

vol - Volume3D to save

store - underlying MINC data type, to be used for storage

Minc2.tfm_to_grid! — Method

tfm_to_grid!(
    tfm::Union{Vector{XFM}, XFM}, 
    grid::AbstractArray{T,4},
    v2w::AffineTransform{C};
    ftol=1.0/80,max_iter=10)::AbstractArray{T,4}

Convert arbitrary transformation into vector field

tfm - transformation to use

grid - output 4D array, will contain vector field

v2w - voxel to world affine transform in the output array

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.tfm_to_grid — Method

tfm_to_grid(
    tfm::Union{Vector{XFM}, XFM},
    ref::G;
    store::Type{T}=Float64,ftol=1.0/80,max_iter=10)::Volume3D{T,4}

Convert arbitrary transformation into Volume3D with 4D array containing vector field

tfm - transformation to use

v2w - voxel to world affine transform in the output array

ftol - tolerance, for inverse transformations

max_iter - maximum number of iterations, for inverse transformations

Minc2.voxel_to_world — Method

voxel_to_world(vol::Volume3D)

Extract voxel to world affine transform from a Volume3D

Minc2.world_to_voxel — Method

world_to_voxel(vol::Volume3D)

Extract world to voxel affine transform from a Volume3D

Minc2.GridTransform — Method

GridTransform(vol::V)

Create GridTransform from Volume3D, assume it is a vector field

vol - Volume3D with 4D array containing vector field

Minc2.InverseGridTransform — Method

InverseGridTransform(vol::V)

Create InverseGridTransform from Volume3D, assume it is a vector field

vol - Volume3D with 4D array containing vector field

Minc2.Volume3D — Type

Volume3D{T,N}

An abstract 3D volume, could be vector field or volume with time dimension

Minc2.Volume3D — Method

Volume3D(vol::Array{T,N}, v2w::AffineTransform{Float64}; 
    history::Union{AbstractString,Nothing}=nothing)::Volume3D{T,N}

Create Volume3D from an array and affine transform

vol - underlying array

v2w - voxel to world affine transform

history - minc history

Minc2.Volume3D — Method

Volume3D(vol::Array{T,N}, like::Volume3D; 
    history::Union{AbstractString,Nothing}=nothing)::Volume3D{T,N}

Create Volume3D from an array and another Volume3D that is used for sampling information

vol - underlying array

like - Volume3D that is used for sampling information

history - minc history

Minc2.lps_to_ras — Method

lps_to_ras(xfm::Minc2.AffineTransform)

Convert coordinates from LPS to RAS

Minc2.lps_to_ras — Method

lps_to_ras(vol::Minc2.Volume3D{T})::Minc2.Volume3D{T}

Convert coordinates from LPS to RAS

Minc2.read_ants_transform — Method

read_ants_transform(fn::AbstractString; store::Type{T}=Float32)::Minc2.AnyTransform

Read .txt and .nii(.nii.gz) transforms produces by ANTs

Minc2.read_itk_nifti_transform — Method

read_itk_nifti_transform(fn::AbstractString;
    store::Type{T}=Float32)::Minc2.GridTransform{Float64,T}

Read ANTs style warp transform

Minc2.read_itk_txt_transform — Method

read_itk_txt_transform(fn::AbstractString)::Minc2.AffineTransform{Float64}

Read ITK legacy transform in .txt format

Minc2.read_nifti_volume — Method

read_nifti_volume(fn::AbstractString; store::Type{T}=Float64)::Volume3D{T}

Read Volume3D from .nii or .nii.gz file

Minc2.save_nifti_volume — Method

save_nifti_volume(fn::AbstractString, vol::Volume3D{T}; 
    store::Type{S}=Float32, history=nothing)

Save Volume3D into .nii or .nii.gz file

Minc2.write_itk_nifti_transform — Method

write_itk_nifti_transform(fn::AbstractString,
    xfm::Minc2.GridTransform{Float64,T}; store::Type{S}=Float32)

Write ANTs style warp transform

Minc2.write_itk_txt_transform — Method

write_itk_txt_transform(fn::AbstractString,
    xfm::Minc2.AffineTransform{Float64})

Write affine transform in ITK .txt format

Minc2.append_grid_transform — Method

append_grid_transform(h::TransformHandle, grid_file::String; inv::Bool=false)

Low level: Append grid transform to an open transformation handle

Minc2.append_linear_transform — Method

append_linear_transform(h::TransformHandle, lin::AffineTransform)

Low level: Append affine transform to an open transformation handle

Minc2.concat_xfm — Method

concat_xfm(h::TransformHandle, i::TransformHandle)

Low level: concatenate two transfomations

Minc2.get_grid_transform — Method

get_grid_transform(h::TransformHandle; n::Int64=0)

Low level: extract reference to a grid file from open handle

Minc2.get_linear_transform — Method

get_linear_transform(h::TransformHandle; n::Int64=0)::AffineTransform{Float64}

Low level: extract AffineTransform{Float64} from open handle

Minc2.get_linear_transform_param — Method

get_linear_transform_param(h::TransformHandle;n::Int64=0,center::Union{Nothing,Vector{Float64}}=nothing)

Low level: extact transformation parameters from affine transform , returns named tuple with fields center, translations, scales, shears, rotations

Minc2.get_n_concat — Method

get_n_concat(h::TransformHandle)::Int64

Low level: Get number of transformations in open handle

Minc2.get_n_type — Method

get_n_type(h::TransformHandle; n::Int64=0)::XFM

Low level: Get transform type for nth transform in open handle

Minc2.inverse_transform_point — Method

inverse_transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Inverse transform point xyz using minc transform h

Minc2.invert_transform — Method

invert_transform(h::TransformHandle)

Low level: Invert minc transform h

Minc2.load_transforms — Method

load_transforms(h::TransformHandle)::Vector{AnyTransform}

Low level: Load all transforms from open .XFM handle

Minc2.load_transforms — Method

load_transforms(fname::String)::Vector{AnyTransform}

Load transformations from .xfm file fname

Minc2.open_xfm_file — Method

Low level: Open transform xfm file, return handle

Minc2.save_transforms — Method

save_transforms(fname::String, 
    xfm::Union{Vector{XFM}, XFM};
    grid_store::Type{T}=Float32 ) where {T, XFM<:AnyTransform}

Save transformations into .xfm file,

Arguments

fname: output file name

xfm: vector of transformations to save, or a single transformation

grid_store: type of storage for grid files (default: Float32)

Minc2.save_xfm_file — Method

save_xfm_file(h::TransformHandle, path::String)

Low level: Save information into file from an open handle

Minc2.transform_point — Method

transform_point(h::TransformHandle, xyz::Vector{Float64})::Vector{Float64}

Low level: Transform point xyz using minc transform h

Minc2.TransformHandle — Type

Low level: minc2_simple XFM transform handle

Minc2.XFM — Type

@enum XFM

Low level: transformation types stored in .xfm file